Invert extension web UI through events
